refactored main routine content into methods
while the main rotine is kept out of obscure code and we
can get the list of tests in our suite in the individual
test case scenarios we still have lots of duplication. if anything
needs to be changed in the future there we would need to apply this to
every single test case.