test scenarios for boundary gear behavior
The gear should not be decreased to idle even upon low RPM
and no suggestion to increase gear to non-existing one should happen when
we are already on the highest gear regardless of how much of RPM we have.
To reflect that we added two new test cases and followed with GearPrompter
necessary adjustments to get them passing.

Notice that max gear is fixed as static field in GearPrompter for now.