scenarios for reverse and idle gears
Again we first write tests and the adjust GearPrompter accordingly.
We expect no gear change in case of rerverse or idle gear regardless
of the rpm in place, but we chose high rpm for test scenarios for now.
Econo rpm would not make much sense since it would be automatically
covered by previous implementation. In the next step we will be adding
low rpm but we will be doing this without introducing new test case
explicltly but rather create parametrized test.