neutral and reverse gear scenarios extended tests
Notice how we now use Theory instead Fact. With that we can now
have givenRpm parameter on our tests which XUnit testing framework
will fill in for us automatically based on InlineData provided.

We are currently duplicating three InlineData attributes on both
scenarios but we could use MemberData to keep this duplication out of
our tests. We will be taking care of that later.