input validation tests
Since we want to guard prompter logic we have implemented a
few scenarios to cover that. In particular no negative RPM is accepted
or gears outside the supported bounds. Validation implementation has been
included in PrompterInput and is being called from GearPrompter itself.

The only parametrization for now is maxGear as the remaining gears are
assumed to be fixed for our implementation. That may change in future
though.