Refactored to cleaner constants for rpms
So far we used rpm constants in the tests directly. They happened to be
all over the same 1000, 2000 and 3000 respectively for low, econo and high
rpm. We also used 1500 for validations where we didn't care for actual
value.

Having those values scattered around all tests make them hard to maintain.
Imagine we need to change those values in future when requirement change.
For example 2000 may no longer be economical RPM for different kind of
engine. Also as we will soon find out the choice we made for low, econo
and high are actually not that good and we will want to update them even
within the current implementation.

To avoid the need to update multiple places and more importantly to
attach meaning that mere constant does not provide we have refactored
those values out to some static fields. As their defintion is not actually
very intetesting to one looking at test they are placed at the bottom
of test. We will be extracting them out of the test class at all soon.

Notice that Theory is now followed with MemberData so that we can pull
from those defined values and have duplication reduced further as
every parametrized test uses the same data source now. Even more important
by introducing property AllRpms we were able to add meaningul name to it,
which we were unable by using InlineData to inject those values directly.