better rpm constants that reflect corner cases
As noted previously we will want to adjust RPMs we used in unit tests.
The previous steps makes it easy to play with those.

The updated values are a bit different from generic previous ones.
They are very specific. We used 2501 for high rpm instead 3000, why?
Please switch high rpm to 3000 back and run all tests. They still pass.
Good? Not necessarily. Keeping 3000 as the high rpm go to GearPrompter
and try playing with the condition that we have for gear increase piece.

Can you change it and still get all tests passing? Now go back to 2501 and try the same.

You have probably tried changing 2500 constant in the code to 2600 or
maybe changed greater than check to greater than or equals. The rules of
thumb is "any meanigful change to the production code should cause
at least one test to fail". The meaningful part is hard for computers but
for humans is pretty easy - we can easily add "true &&" at the beginning
of the condition and still make all tests passed even with the 2501 value.
There are frameworks that attempt to create such mutations automatically
to help assessing how good your tests are - you can learn more about the
topic here: https://en.wikipedia.org/wiki/Mutation_testing