externalized gear box specifics out from prompter
One of the most important benefits of having unit tests suite is
that they provide a safety net for production code refactorings and
requirements changes. With obscure code bases with complex code any change
is hard and gets harder with time. This is where unit tests come with
help, but you need to have them to benefit from that obviously.

With safety net of unit tests any refactoring or code change, can be
checked automatically by running the suite assuring nothing broke
elsewhere. Manual tests can catch those as well, but since they are
handled by humans they are error prone themselves. And sloooow compared
to automated ones.

There are some troubles to look for though. Check this commit. We have
decided to get gear box specification outside prompter. This may be mere
refactoring or maybe it was dictated due to requirement to have prompter
to support multiple gear boxes it can be deployed with.

We have changed the constructor to inject those details since they will
not be changing thorugh the lifetime of the prompter.

But this made ALL our tests broken. They actually didn't compile in the
first place due to lack of input parameter on the constructor. Also fixing
that produced bunch of additional code that obscured the scenarios and
this is the most important thing to avoid to keepo maintainable tests
suite.

Note how Validate_NegativeRpm test is now affected by the need to create
some GearboxInfo which doesn't really matter for that test. To avoid that
and make changing prompter code easier in future we will be introducing
ClassFixture pattern in the next step.