moved class fixture creation into test constructor
With XUnit framework individual test class instance are created for each
Fact or Theory case. Thanks to that we can further clean up some tests by
moving fixture instantiation to test class construction.

With other frameworks like NUnit this is not the case and for those
special methods need to be created (marked with [SetUp] attribute).

Notice how most of test cases we have just use fixture without additional
need to customize it. However Validation test case scenario customizes
fixture with custom max gear value to make it explicit what the max gear
is for that case. While it is set to be exactly as the default one it may
not always be the case.