included prompter input creation in the fixture
Our test fixture can also support act and assert parts of our test cases.
Lets start with act first. We need to create PrompterInput to trigger our
recommendation. But for some cases we only care about RPM and for others
only for current gear. This is especially the case for validation logic we
are checking. We tried to handle this via SomeRpm constant so far, but we
can do better. We customize our fixture with current values for RPM and
gear and leave input creation to the fixture itself. If we wish to change
the way we aquire those in future (e.g. from external system hidden behind
some interface) we will have easier job of refactoring, since our tests
again will not need to change.