gear suggestion auditor feature
Now imagine new requirement comes in that requires each gear suggestion to
be audited. This may be due to various of reasons, one being future
analysis to make the implementation better or even adjust the suggestions
based on the usage scenarios.

Notice how having fixture already in place makes our live easier. First we
do not need to touch any of existing tests, even though we are adding new
constructor dependency.

Moreover notice the fact we are changing GearPrompter implementation
significantly. We had multiple return points from this method. It is in
general something we want to avoid, but it was working well till now.
However with the need to audit each suggestion we really do not want to
call auditor twice in our Recommend method.

At this point if we had no test (and most probably much more complex code)
we could worry that our change may introduce some bug. But with
comprehensive test suite we have, we can apply this non-essential (from
business point of view) change with full confidence.

Finally notice we implemented very simple test for auditing. We only check
on count. This is area to improve. Some more tests for that feature are
left as excerice for yourself. Notice that current CountingAuditor
implementation does not store any of audited entries, so it would need to
be extended to allow more elaborate checks.